TST: test_genericmotor.py reorganization.
