+++ tests/unit/test_genericmotor.py
@@ -16,21 +16,6 @@ dry_mass = 8.0
 dry_inertia = (0.2, 0.2, 0.08)
 
 
-@patch("matplotlib.pyplot.show")
-def test_generic_motor_info(mock_show, generic_motor):
-    """Tests the GenericMotor.all_info() method.
-
-    Parameters
-    ----------
-    mock_show : mock
-        Mock of the matplotlib.pyplot.show function.
-    generic_motor : rocketpy.GenericMotor
-        The GenericMotor object to be used in the tests.
-    """
-    assert generic_motor.info() == None
-    assert generic_motor.all_info() == None
-
-
 def test_generic_motor_basic_parameters(generic_motor):
     """Tests the GenericMotor class construction parameters.
